+++ src/main/java/org/ohdsi/webapi/annotation/answer/AnswerRepository.java
@@ -3,4 +3,4 @@ package org.ohdsi.webapi.annotation.answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.ohdsi.webapi.annotation.answer.Answer;
 
-public interface AnswerRepository extends JpaRepository<Answer, Long> {}
+public interface AnswerRepository extends JpaRepository<Answer, Integer> {}
